@@ -2,8 +2,9 @@
   "name": "unoapi-cloud",
   "version": "1.15.47",
   "description": "Unoapi Cloud",
-  "main": "./dist/index.js",
+  "exports": "./dist/index.js",
   "types": "./dist/index.d.ts",
+  "type": "module",
   "files": [
     "./dist"
   ],
@@ -12,8 +13,8 @@
   "license": "GPL-3.0",
   "private": false,
   "scripts": {
-    "lint": "eslint src",
-    "test": "jest --coverage",
+    "lint": "eslint",
+    "test": "node_modules/jest/bin/jest.js --coverage",
     "build": "./node_modules/typescript/bin/tsc -p .",
     "format": "prettier --ignore-path .gitignore --write \"**/*.+(js|ts|json)\"",
     "dev": "./node_modules/nodemon/bin/nodemon.js src/index.ts",
@@ -28,57 +29,59 @@
     "clean:data": "rm -rf ./data/medias/* ./data/sessions/* ./data/stores/*.json"
   },
   "devDependencies": {
-    "@types/express": "^4.17.17",
-    "@types/jest": "^29.4.0",
-    "@types/mime-types": "^2.1.1",
-    "@types/node": "^18.14.1",
+    "@eslint/js": "^9.1.1",
+    "@types/express": "^4.17.21",
+    "@types/jest": "^29.5.12",
+    "@types/mime-types": "^2.1.4",
+    "@types/node": "^20.12.7",
     "@types/node-fetch": "2",
-    "@types/qrcode": "^1.5.0",
-    "@types/supertest": "^2.0.12",
-    "@types/uuid": "^9.0.1",
-    "@types/vcf": "^2.0.3",
-    "@types/ws": "^8.5.4",
-    "@typescript-eslint/eslint-plugin": "^5.53.0",
+    "@types/qrcode": "^1.5.5",
+    "@types/supertest": "^6.0.2",
+    "@types/uuid": "^9.0.8",
+    "@types/vcf": "^2.0.7",
+    "@types/ws": "^8.5.10",
     "@typescript-eslint/parser": "^5.53.0",
+    "@typescript-eslint/eslint-plugin": "^5.53.0",
     "eslint": "^8.34.0",
     "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "jest": "^29.4.3",
+    "jest": "^29.7.0",
     "jest-fetch-mock": "^3.0.3",
-    "jest-mock-extended": "^3.0.3",
-    "nodemon": "^2.0.20",
-    "pino-pretty": "^9.4.0",
-    "prettier": "^2.8.4",
-    "supertest": "^6.3.3",
-    "ts-jest": "^29.0.5",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.0.4"
+    "jest-mock-extended": "^3.0.6",
+    "nodemon": "^3.1.0",
+    "pino-pretty": "^7.0.0",
+    "prettier": "^3.2.5",
+    "supertest": "^7.0.0",
+    "ts-jest": "^29.1.2",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.5",
+    "typescript-eslint": "^7.8.0"
   },
   "dependencies": {
-    "@aws-sdk/client-s3": "^3.341.0",
-    "@aws-sdk/s3-request-presigner": "^3.341.0",
-    "@google-cloud/text-to-speech": "^4.2.2",
-    "@redis/client": "^1.5.6",
+    "@aws-sdk/client-s3": "^3.565.0",
+    "@aws-sdk/s3-request-presigner": "^3.565.0",
+    "@google-cloud/text-to-speech": "^5.2.0",
+    "@redis/client": "^1.5.14",
     "@whiskeysockets/baileys": "^6.7.1",
-    "amqplib": "^0.10.3",
-    "awesome-phonenumber": "^6.2.0",
-    "dotenv": "^16.0.3",
-    "express": "^4.18.2",
-    "jimp": "^0.22.10",
-    "jschardet": "^3.0.0",
-    "link-preview-js": "^3.0.4",
+    "amqplib": "^0.10.4",
+    "awesome-phonenumber": "^6.8.0",
+    "dotenv": "^16.4.5",
+    "express": "^4.19.2",
+    "jimp": "^0.22.12",
+    "jschardet": "^3.1.2",
+    "link-preview-js": "^3.0.5",
     "mime-types": "^2.1.35",
     "node-cache": "^5.1.2",
     "node-fetch": "2",
-    "node-xlsx": "^0.23.0",
-    "pino": "^8.14.1",
-    "qrcode": "^1.5.1",
+    "node-xlsx": "^0.24.0",
+    "pino": "^7.0.0",
+    "qrcode": "^1.5.3",
     "qrcode-terminal": "^0.12.0",
-    "socket.io": "^4.7.4",
-    "socks-proxy-agent": "^8.0.2",
-    "uuid": "^9.0.0",
-    "vcf": "^2.1.1",
+    "socket.io": "^4.7.5",
+    "socks-proxy-agent": "^8.0.3",
+    "uuid": "^9.0.1",
+    "vcf": "^2.1.2",
     "xlsx": "^0.18.5",
-    "yaml": "^2.2.1"
+    "yaml": "^2.4.2"
   }
 }
